requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
apache-airflow==2.9.1
apache-airflow-providers-common-io==1.3.2
apache-airflow-providers-common-sql==1.14.0
apache-airflow-providers-fab==1.1.1
apache-airflow-providers-ftp==3.9.1
apache-airflow-providers-google==10.19.0
apache-airflow-providers-http==4.11.1
apache-airflow-providers-imap==3.6.1
apache-airflow-providers-smtp==1.7.1
apache-airflow-providers-sqlite==3.8.1
apache-beam==2.56.0
google-cloud-bigquery<3.21.0
apispec==6.6.1
argcomplete==3.4.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
black==24.4.2
blinker==1.8.2
cachelib==0.9.0
cachetools==5.3.3
cattrs==23.2.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
cloudpickle==2.2.1
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
crcmod==1.7
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
db-dtypes==1.2.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
fastavro==1.9.4
fasteners==0.19
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.6.0
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==7.1.0
gcloud-aio-storage==9.2.0
gcsfs==2024.6.0
google-ads==24.1.0
google-analytics-admin==0.22.7
google-api-core==2.19.0
google-api-python-client==2.134.0
google-apitools==0.5.31
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.56.0
google-cloud-appengine-logging==1.4.3
google-cloud-audit-log==0.2.5
google-cloud-automl==2.13.3
google-cloud-batch==0.17.21
google-cloud-bigquery==3.20.1
google-cloud-bigquery-datatransfer==3.15.3
google-cloud-bigquery-storage==2.25.0
google-cloud-bigtable==2.23.1
google-cloud-build==3.24.0
google-cloud-compute==1.19.0
google-cloud-container==2.47.0
google-cloud-core==2.4.1
google-cloud-datacatalog==3.19.0
google-cloud-dataflow-client==0.8.10
google-cloud-dataform==0.5.9
google-cloud-dataplex==2.0.0
google-cloud-dataproc==5.9.3
google-cloud-dataproc-metastore==1.15.3
google-cloud-datastore==2.19.0
google-cloud-dlp==3.16.0
google-cloud-kms==2.23.0
google-cloud-language==2.13.3
google-cloud-logging==3.10.0
google-cloud-memcache==1.9.3
google-cloud-monitoring==2.21.0
google-cloud-orchestration-airflow==1.12.1
google-cloud-os-login==2.14.3
google-cloud-pubsub==2.21.1
google-cloud-pubsublite==1.10.0
google-cloud-recommendations-ai==0.10.10
google-cloud-redis==2.15.3
google-cloud-resource-manager==1.12.3
google-cloud-run==0.10.5
google-cloud-secret-manager==2.20.0
google-cloud-spanner==3.47.0
google-cloud-speech==2.26.0
google-cloud-storage==2.16.0
google-cloud-storage-transfer==1.11.3
google-cloud-tasks==2.16.3
google-cloud-texttospeech==2.16.3
google-cloud-translate==3.15.3
google-cloud-videointelligence==2.13.3
google-cloud-vision==3.7.2
google-cloud-workflows==1.14.3
google-crc32c==1.5.0
google-re2==1.1.20240601
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpc-interceptor==0.15.4
grpcio==1.64.1
grpcio-gcp==0.2.2
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Js2Py==0.74
json-merge-patch==0.2
jsonpickle==3.2.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
limits==3.12.0
linkify-it-py==2.0.3
lockfile==0.12.2
looker-sdk==24.10.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.1
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.7.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
orjson==3.10.5
overrides==7.7.0
packaging==24.1
pandas==2.1.4
pandas-gbq==0.23.1
pathspec==0.12.1
pendulum==3.0.0
platformdirs==4.2.2
pluggy==1.5.0
prison==0.2.1
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
py4j==0.10.9.7
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydata-google-auth==1.8.2
pydot==1.4.2
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.8.0
pymongo==4.7.3
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyspark==3.5.1
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.18.1
rsa==4.9
ruff==0.4.10
setproctitle==1.3.3
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
sqlalchemy-bigquery==1.11.0
SQLAlchemy-JSONField==1.0.2
sqlalchemy-spanner==1.7.0
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
tabulate==0.9.0
tenacity==8.4.1
termcolor==2.4.0
text-unidecode==1.3
time-machine==2.14.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==2.2.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.19.2
zstandard==0.22.0
